+++ pangea-sdk/service/ai_guard/api.go
@@ -37,6 +37,15 @@ func (e *aiGuard) Guard(ctx context.Context, body GuardRequest) (*pangea.PangeaR
 	return request.DoPost(ctx, e.Client, "v1beta/guard", &body, &GuardResult{})
 }
 
+// @summary Guard LLM input and output
+//
+// @description Analyze and redact content to avoid manipulation of the model, addition of malicious content, and other undesirable data transfers.
+//
+// @operationId ai_guard_post_v1beta_guard_async
+func (e *aiGuard) GuardAsync(ctx context.Context, body GuardRequest) (*pangea.PangeaResponse[GuardResult], error) {
+	return request.DoPost(ctx, e.Client, "v1beta/guard_async", &body, &GuardResult{})
+}
+
 // @operationId ai_guard_post_v1beta_config
 func (e *aiGuard) GetServiceConfig(ctx context.Context, body GetServiceConfigParams) (*pangea.PangeaResponse[ServiceConfig], error) {
 	return request.DoPost(ctx, e.Client, "v1beta/config", &body, &ServiceConfig{})
@@ -417,21 +426,24 @@ type MultimodalMessage struct {
 type GuardRequest struct {
 	pangea.BaseRequest
 
-	Messages           []MultimodalMessage `json:"messages,omitempty"` // Prompt content and role array in JSON format. The `content` is the multimodal text or image input that will be analyzed.
-	Recipe             string              `json:"recipe,omitempty"`   // Recipe key of a configuration of data types and settings defined in the Pangea User Console. It specifies the rules that are to be applied to the text, such as defang malicious URLs.
-	Debug              bool                `json:"debug,omitempty"`    // Setting this value to true will provide a detailed analysis of the text data
-	Overrides          Overrides           `json:"overrides,omitempty"`
-	AppName            string              `json:"app_name,omitempty"`             // Name of source application.
-	LlmProvider        string              `json:"llm_provider,omitempty"`         // Underlying LLM.  Example: 'OpenAI'.
-	Model              string              `json:"model,omitempty"`                // Model used to perform the event. Example: 'gpt'.
-	ModelVersion       string              `json:"model_version,omitempty"`        // Model version used to perform the event. Example: '3.5'.
-	RequestTokenCount  int32               `json:"request_token_count,omitempty"`  // Number of tokens in the request.
-	ResponseTokenCount int32               `json:"response_token_count,omitempty"` // Number of tokens in the response.
-	SourceIp           string              `json:"source_ip,omitempty"`            // IP address of user or app or agent.
-	SourceLocation     string              `json:"source_location,omitempty"`      // Location of user or app or agent.
-	TenantId           string              `json:"tenant_id,omitempty"`            // For gateway-like integrations with multi-tenant support.
-	SensorMode         string              `json:"sensor_mode,omitempty"`          // (AIDR) sensor mode.
-	Context            map[string]any      `json:"context,omitempty"`              // (AIDR) Logging schema.
+	// 'messages' (required) contains Prompt content and role array in JSON
+	// format. The `content` is the multimodal text or image input that will be
+	// analyzed. Additional properties such as 'tools' may be provided for analysis.
+	Input              any            `json:"input,omitempty"`
+	Recipe             string         `json:"recipe,omitempty"` // Recipe key of a configuration of data types and settings defined in the Pangea User Console. It specifies the rules that are to be applied to the text, such as defang malicious URLs.
+	Debug              bool           `json:"debug,omitempty"`  // Setting this value to true will provide a detailed analysis of the text data
+	Overrides          Overrides      `json:"overrides,omitempty"`
+	AppName            string         `json:"app_name,omitempty"`             // Name of source application.
+	LlmProvider        string         `json:"llm_provider,omitempty"`         // Underlying LLM.  Example: 'OpenAI'.
+	Model              string         `json:"model,omitempty"`                // Model used to perform the event. Example: 'gpt'.
+	ModelVersion       string         `json:"model_version,omitempty"`        // Model version used to perform the event. Example: '3.5'.
+	RequestTokenCount  int32          `json:"request_token_count,omitempty"`  // Number of tokens in the request.
+	ResponseTokenCount int32          `json:"response_token_count,omitempty"` // Number of tokens in the response.
+	SourceIp           string         `json:"source_ip,omitempty"`            // IP address of user or app or agent.
+	SourceLocation     string         `json:"source_location,omitempty"`      // Location of user or app or agent.
+	TenantId           string         `json:"tenant_id,omitempty"`            // For gateway-like integrations with multi-tenant support.
+	SensorMode         string         `json:"sensor_mode,omitempty"`          // (AIDR) sensor mode.
+	Context            map[string]any `json:"context,omitempty"`              // (AIDR) Logging schema.
 }
 
 type GuardResult struct {