+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - AI Guard: multimodal guard.
+- AI Guard: async guard.
 
 ## 5.3.0-beta.2 - 2025-05-22