+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.3.0-beta.3 - 2025-09-02
+
 ### Added
 
 - AI Guard: multimodal guard.